cleanup -- name changed to ferrox
