+++ Source/ElectrostaticSolver.H
@@ -4,10 +4,7 @@
 #include "FerroX.H"
 #include "Input/BoundaryConditions/BoundaryConditions.H"
 #include "Input/GeometryProperties/GeometryProperties.H"
-#include "Code.H"
-//#include "Utils/SelectWarpXUtils/WarpXUtil.H"
-//#include "Utils/SelectWarpXUtils/WarpXProfilerWrapper.H"
-//#include "Utils/CodeUtils/CodeUtil.H"
+//#include "Code.H"
 
 using namespace amrex;
 using namespace FerroX;
@@ -43,8 +40,8 @@ void ComputePoissonRHS_Newton(MultiFab& PoissonRHS,
                               MultiFab& PoissonPhi, 
                               MultiFab& alpha_cc);
 
-void SetPoissonBC(c_Code& rCode, std::array<std::array<amrex::LinOpBCType,AMREX_SPACEDIM>,2>& LinOpBCType_2d, bool& all_homogeneous_boundaries, bool& some_functionbased_inhomogeneous_boundaries, bool& some_constant_inhomogeneous_boundaries);
+void SetPoissonBC(c_FerroX& rFerroX, std::array<std::array<amrex::LinOpBCType,AMREX_SPACEDIM>,2>& LinOpBCType_2d, bool& all_homogeneous_boundaries, bool& some_functionbased_inhomogeneous_boundaries, bool& some_constant_inhomogeneous_boundaries);
 
-void Fill_Constant_Inhomogeneous_Boundaries(c_Code& rCode, MultiFab& PoissonPhi);
-void Fill_FunctionBased_Inhomogeneous_Boundaries(c_Code& rCode, MultiFab& PoissonPhi);
+void Fill_Constant_Inhomogeneous_Boundaries(c_FerroX& rFerroX, MultiFab& PoissonPhi);
+void Fill_FunctionBased_Inhomogeneous_Boundaries(c_FerroX& rFerroX, MultiFab& PoissonPhi);
