--- Source/Input/MacroscopicProperties/MacroscopicProperties_fwd.H
@@ -1,7 +1,0 @@
-#ifndef MACROSCOPIC_PROPERTIES_FWD_H
-#define MACROSCOPIC_PROPERTIES_FWD_H
-
-class c_MacroscopicProperties;
-
-#endif 
-
