+++ Source/Make.package
@@ -4,10 +4,8 @@ CEXE_sources += Initialization.cpp
 CEXE_sources += ChargeDensity.cpp
 CEXE_sources += TotalEnergyDensity.cpp
 CEXE_sources += FerroX.cpp
-CEXE_sources += Code.cpp
 
 CEXE_headers += FerroX.H
-CEXE_headers += Code.H
 CEXE_headers += ElectrostaticSolver.H
 CEXE_headers += Initialization.H
 CEXE_headers += ChargeDensity.H
