+++ Source/Utils/FerroXUtils/FerroXUtil.cpp
@@ -1,7 +1,4 @@
-#include <CodeUtil.H>
-//#include "../../Code.H"
-//#include "../../Input/MacroscopicProperties/MacroscopicProperties.H"
-//#include "../../PostProcessor/PostProcessor.H"
+#include <FerroXUtil.H>
 
 using namespace amrex;
